requirements.txt
python-constraint==1.4.0
pandas==1.4.4
streamlit==1.20.0